2015/06 (!) Probably a Fire Hazard, p1 solved
Used booleans for the values, with the naive hope that the language would
detect a bit array and automatically optimize for space. How silly of
me. Should try if numpy helps. Anyway, a mere million entries is not worth
optimizing for space even on a 32 bit machine.

So I should have used 0, 1 for values.

input.split("\n") produces an empty string as the last element, so
input.strip().split("\n") seems a reasonable fix.

Comparing the verb in the inner loop is slow.

Turning on and off one by one is slow, should try replacing a slice.